lab_8_SQL_FULL_ACCESS_PY/mcp_server/requirements.txt
# MCP server (FastMCP + Postgres)
mcp[cli]>=1.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
